🔧 (appsettings.json, TfsTeamProjectAuthentication.cs, TfsEndpointOptions.cs, TfsWorkItemMigrationProcessorOptions.cs, EndpointOptions.cs, ProcessorOptions.cs): add new configuration fields and improve JSON serialization handling
Add the "ReflectedWorkItemIdField" to multiple sections in appsettings.json to support custom work item ID fields. Improve JSON serialization handling by adding `NullValueHandling.Ignore` and `DefaultValueHandling.Ignore` attributes to various properties in the codebase. This ensures that default values and null properties are not serialized, reducing the size of the configuration files and making them cleaner.

These changes enhance the flexibility and maintainability of the configuration files, allowing for more customized and efficient setups.